% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{assign_cells}
\alias{assign_cells}
\title{Assign cells to a cell type based on maximum marker scores.}
\usage{
assign_cells(scores, group_assignment = NULL)
}
\arguments{
\item{scores}{A marker score matrix as returned by score_cells.}

\item{group_assignment}{Character vector specifying to which group (or superclass)
each cell has been assigned when performing hierarchical assignment. The cell
can then only be assigned a cell type within this group. If no vector is provided,
any cell can be assigned to any cell type. For this option to work, score_cells must be run
on a MetaMarkers marker table where groups have been specified.}
}
\value{
A data.frame containing one row per cell, showing its group, its
assigned cell type, the marker score for that cell type and the marker
enrichment for that cell type.
}
\description{
Assign cells to a cell type based on maximum marker scores.
}
