% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{compute_marker_enrichment}
\alias{compute_marker_enrichment}
\title{Compute marker enrichment (score over expected score, under the hypothesis
that all marker sets are expressed equally in the cell).}
\usage{
compute_marker_enrichment(scores, by_group = TRUE)
}
\arguments{
\item{scores}{A marker score matrix as returned by score_cells.}

\item{by_group}{Boolean. If TRUE, the expected score is computed based
on neighboring cell types. For this option to work, score_cells must be run
on a MetaMarkers marker table where groups have been specified.}
}
\value{
A cell type x cell matrix containing marker enrichment scores.
}
\description{
Compute marker enrichment (score over expected score, under the hypothesis
that all marker sets are expressed equally in the cell).
}
