% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{compute_markers}
\alias{compute_markers}
\title{Compute differential expression statistics for given dataset and cell types,
stratified by groups.}
\usage{
compute_markers(
  expression,
  cell_type_labels,
  group_labels = rep("all", length(cell_type_labels)),
  two_tailed = FALSE,
  tie_correction = FALSE,
  genes_are_rows = TRUE
)
}
\arguments{
\item{expression}{Expression matrix (may be sparse).}

\item{cell_type_labels}{Character vector providing
cell type names for each sample in the expression matrix.}

\item{group_labels}{Character vector providing hierarchical grouping for
cell types (one group name for each sample in the expression matrix).}

\item{two_tailed}{Boolean. If FALSE, only upregulated genes are considered
significant (ROC test).}

\item{tie_correction}{Boolean. For the ROC test, should tie correction be
applied? Note that skipping tie correction is slightly conservative.}

\item{genes_are_rows}{Boolean. In the expression matrix, were genes provided
as rows (as in SingleCellExperiment or Seurat objects)?}
}
\value{
A tibble containing basic differential expression statistics for
all cell types and genes. All statistics are 1-vs-rest within groups. Note
genes with duplicate names will be removed.
}
\description{
Compute differential expression statistics for given dataset and cell types,
stratified by groups.
}
