% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{compute_umap}
\alias{compute_umap}
\title{Compute UMAP coordinates for a target dataset.}
\usage{
compute_umap(expr)
}
\arguments{
\item{expr}{Expression matrix (with genes as rows) of the target dataset in
sparse or dense format. We recommend providing log-normalized counts for
carefully selected highly variable genes.}
}
\value{
A tibble with one row per cell and 3 columns containing the sample
name, UMAP coordinate 1 and UMAP coordinate 2.
}
\description{
Compute UMAP coordinates for a target dataset.
}
