% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{convert_to_cpm}
\alias{convert_to_cpm}
\title{Convert expression matrix to CPM (library size normalization).}
\usage{
convert_to_cpm(M)
}
\arguments{
\item{M}{Expression matrix (with genes as rows) in sparse or dense format.}
}
\value{
Matrix in same format as input.
}
\description{
Convert expression matrix to CPM (library size normalization).
}
