% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurocs.R
\name{design_matrix}
\alias{design_matrix}
\title{Convert character vector to one-hot encoded matrix, with one column per
unique string.}
\usage{
design_matrix(cell_type, scale_columns = FALSE)
}
\arguments{
\item{cell_type}{Character vector.}

\item{scale_columns}{Boolean, should columns be normalized to sum to 1?}
}
\description{
Convert character vector to one-hot encoded matrix, with one column per
unique string.
}
