% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{export_markers}
\alias{export_markers}
\title{Export marker statistics to file.}
\usage{
export_markers(de_stats, filename, gzip = TRUE)
}
\arguments{
\item{de_stats}{Tibble obtained by calling \code{compute_de_stats}.}

\item{filename}{File name.}

\item{gzip}{Boolean. Should output file be zipped to spare memory?}
}
\description{
This function will export marker statistics in a simple format, including a short
summary header with cell types analyzed and date of export.
}
