% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{export_markers_by_group}
\alias{export_markers_by_group}
\title{Export markers statistics to file, one file per group.}
\usage{
export_markers_by_group(de_stats, output_prefix, gzip = TRUE)
}
\arguments{
\item{de_stats}{Tibble obtained by calling \code{compute_markers}.}

\item{output_prefix}{File prefix, typically a path or a path + prefix.
Each group will be exported to a standardized file name, consisting of
output_prefix, a standardized version of the group name
using \code{make.names}) and a .csv extension.}

\item{gzip}{Boolean. Should output file be zipped to spare memory?}
}
\description{
This function will export marker statistics in a simple format, including a short
summary header with cell types analyzed and date of export.
}
