% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{export_markers_for_cell_type}
\alias{export_markers_for_cell_type}
\title{Export marker statistics of given cell type to file.}
\usage{
export_markers_for_cell_type(de_stats, cell_type, filename, gzip = TRUE)
}
\arguments{
\item{de_stats}{Tibble obtained by calling \code{compute_markers}.}

\item{cell_type}{Name of cell type of interest (contained in de_stats).}

\item{filename}{File name.}

\item{gzip}{Boolean. Should output file be zipped to spare memory?}
}
\description{
This function will export marker statistics in a simple format, including a short
summary header with cell types analyzed and date of export.
}
