% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_markers.R
\name{export_meta_markers}
\alias{export_meta_markers}
\title{Export meta-analytic markers to file (including header with time stamp).}
\usage{
export_meta_markers(
  meta_markers,
  filename,
  dataset_names,
  order_by = "auroc",
  fdr_threshold = 0.05,
  fc_threshold = 4,
  detection_threshold = 0,
  gzip = TRUE
)
}
\arguments{
\item{meta_markers}{Meta-marker table obtained with \code{make_meta_marker}.}

\item{filename}{File where to write meta-marker table.}

\item{dataset_names}{Names of datasets used to compute meta-markers.}

\item{order_by}{Secondary statistic by which meta-markers have been ranked.}

\item{fdr_threshold}{FDR threshold used to identify Differentially
Expressed (DE) genes.}

\item{fc_threshold}{Fold change threshold used to identify DE genes.}

\item{detection_threshold}{Detection rate threshold used to identify DE genes.}

\item{gzip}{Boolean. Should output file be zipped to spare memory?}
}
\description{
Export meta-analytic markers to file (including header with time stamp).
}
