% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_cell_type}
\alias{get_cell_type}
\title{Extract cell type information from a cell type label.}
\usage{
get_cell_type(cell_type_label)
}
\arguments{
\item{cell_type_label}{Character vector with strings in "group|cell_type" format.}
}
\value{
Character vector containing only the "cell_type" part. If any of the input strings,
is not in "group|cell_type" format, the function simply returns the input.
}
\description{
Extract cell type information from a cell type label.
}
