% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_group}
\alias{get_group}
\title{Extract group information from a cell type label.}
\usage{
get_group(cell_type_label)
}
\arguments{
\item{cell_type_label}{Character vector with strings in "group|cell_type" format.}
}
\value{
Character vector containing only the "group" part. If any of the input strings,
is not in "group|cell_type" format, the function returns a vector of empty strings.
}
\description{
Extract group information from a cell type label.
}
