% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_markers.R
\name{make_meta_markers}
\alias{make_meta_markers}
\title{Extract meta-analytic markers from multiple datasets.}
\usage{
make_meta_markers(
  marker_lists,
  order_by = "auroc",
  fdr_threshold = 0.05,
  fc_threshold = 4,
  detection_threshold = 0,
  detailed_stats = FALSE,
  common_genes_only = TRUE
)
}
\arguments{
\item{marker_lists}{Named list, where elements are marker statistics obtained
with \code{compute_markers} and names are names of the dataset from which markers
have been extracted.}

\item{order_by}{Secondary statistic by which meta-markers should be ranked.}

\item{fdr_threshold}{FDR threshold for a gene to be considered Differentially
Expressed (DE).}

\item{fc_threshold}{Fold change threshold for a gene to be DE.}

\item{detection_threshold}{Detection rate threshold for a gene to be DE.}

\item{detailed_stats}{Boolean. By default, only output a list of best markers,
alternatively output additional statistics, such as average AUROC, FC, etc.}

\item{common_genes_only}{Boolean. Keep only genes that are common to all
datasets?}
}
\value{
A tibble containing ranked meta-markers and (if desired) average DE
statistics for each cell type. Within each cell type, meta-markers are
ranked by recurrence (# datasets in which gene is DE), then by a secondary
statistics, as specified in "order_by" (AUROC by default).
}
\description{
Extract meta-analytic markers from multiple datasets.
}
