% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_assignments}
\alias{plot_assignments}
\title{Plot cell type assignments in UMAP space.}
\usage{
plot_assignments(
  assignments,
  umap_coordinates,
  enrichment_threshold = 1,
  point_size = 1,
  rasterize_umap = FALSE
)
}
\arguments{
\item{assignments}{A cell type assignment table as returned by assign_cells.}

\item{umap_coordinates}{A (cell x 2) data.frame containing UMAP coordinates, where the ordering
of cells is the same as in the score matrix}

\item{enrichment_threshold}{Numeric value defining a threshold below which a
cell should be considered unassigned (QC metric). 1 by default.}

\item{point_size}{Numeric value indicating the dot size.}

\item{rasterize_umap}{Boolean. If TRUE, the UMAP part of the figure is rasterized.
This is particularly useful for large datasets.}
}
\value{
A ggplot2 object.
}
\description{
Plot cell type assignments in UMAP space.
}
