% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_front.R
\name{plot_marker_expression}
\alias{plot_marker_expression}
\title{Plot expression of markers in cell type as violin plots.}
\usage{
plot_marker_expression(expression_matrix, genes, cell_type_label)
}
\arguments{
\item{expression_matrix}{Expression matrix (with genes as rows) in sparse or
dense format.}

\item{genes}{Vector of gene names.}

\item{cell_type_label}{Vector of cell type names (must be same length as
number of columns in expression_matrix)}
}
\value{
ggplot2 object
}
\description{
Plot expression of markers in cell type as violin plots.
}
