% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_marker_scores}
\alias{plot_marker_scores}
\title{Plot markers scores in UMAP space (one facet per marker set).}
\usage{
plot_marker_scores(
  scores,
  umap_coordinates,
  normalize_scores = FALSE,
  point_size = 1,
  rasterize_umap = FALSE
)
}
\arguments{
\item{scores}{A marker score matrix as returned by score_cells or compute_marker_enrichment.}

\item{umap_coordinates}{A (cell x 2) data.frame containing UMAP coordinates, where the ordering
of cells is the same as in the score matrix}

\item{normalize_scores}{Boolean. If TRUE, marker scores are renormalized between
0 and 1 for all marker sets to obtain similar color ranges across facets.}

\item{point_size}{Numeric value indicating the dot size.}

\item{rasterize_umap}{Boolean. If TRUE, the UMAP part of the figure is rasterized.
This is particularly useful for large datasets.}
}
\value{
A ggplot2 object.
}
\description{
Plot markers scores in UMAP space (one facet per marker set).
}
