% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_front.R
\name{plot_pareto_markers}
\alias{plot_pareto_markers}
\title{Plot best markers in the (fold change of detection, AUROC) space.}
\usage{
plot_pareto_markers(
  meta_markers,
  cell_type_name,
  min_recurrence = 1,
  min_auroc = 0.5,
  min_fc = 0,
  fc_threshold = 3,
  auroc_threshold = 0.8
)
}
\arguments{
\item{meta_markers}{Meta-marker table obtained with \code{make_meta_marker}.}

\item{cell_type_name}{Name of cell type to be plotted (must be contained
in meta-markers).}

\item{min_recurrence}{Recurrence threshold for a marker to be plotted
(# of datasets where marker was DE)}

\item{min_auroc}{AUROC threshold for a marker to be plotted.}

\item{min_fc}{Fold change of detection threshold for a marker to be plotted.}

\item{fc_threshold}{Threshold at which to draw a dashed line on the
AUROC axis (set to NULL to skip).}

\item{auroc_threshold}{Threshold at which to draw a dashed line on the
fold change axis (set to NULL to skip).}
}
\description{
This visualization displays markers meeting minimal threshold conditions
in the (fold change of detection, AUROC) space. The best markers
(Pareto front) are highlighted and labeled.
}
