% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_front.R
\name{plot_pareto_summary}
\alias{plot_pareto_summary}
\title{Plot Pareto fronts (best markers) of all cell types in the (fold change of detection, AUROC) space.}
\usage{
plot_pareto_summary(meta_markers)
}
\arguments{
\item{meta_markers}{Meta-marker table obtained with \code{make_meta_marker}.}
}
\description{
Plot Pareto fronts (best markers) of all cell types in the (fold change of detection, AUROC) space.
}
