% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{read_markers}
\alias{read_markers}
\title{Import marker statistics.}
\usage{
read_markers(filename, header_size = 1)
}
\arguments{
\item{filename}{File name.}

\item{header_size}{Size of header (# rows) containing meta-information such
as time stamp and list of cell types (for compatibility purposes only).}
}
\description{
Import marker statistics.
}
