% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_markers.R
\name{read_meta_markers}
\alias{read_meta_markers}
\title{Import meta-marker statistics.}
\usage{
read_meta_markers(filename, header_size = 1)
}
\arguments{
\item{filename}{File name.}

\item{header_size}{Size of header (# rows) containing meta-information such
as time stamp and list of dataset names (for compatibility purposes only).}
}
\description{
Import meta-marker statistics.
}
