% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{remove_duplicated_genes}
\alias{remove_duplicated_genes}
\title{Remove duplicate gene from a marker table (individual dataset)}
\usage{
remove_duplicated_genes(marker_stats)
}
\arguments{
\item{marker_stats}{Tibble obtained by calling \code{compute_marker_stats}.}
}
\value{
Tibble with marker stats, all duplicated genes being removed.
}
\description{
Remove duplicate gene from a marker table (individual dataset)
}
