% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{score_cells}
\alias{score_cells}
\title{Compute marker scores (average marker expression) in a target dataset}
\usage{
score_cells(expr, marker_set)
}
\arguments{
\item{expr}{Expression matrix (with genes as rows) of the target dataset in
sparse or dense format. We recommend providing log-normalized counts.}

\item{marker_set}{Marker sets, usually a marker data.frame generated by
MetaMarkers. Other possible inputs are: a named list (where names
are cell types and values are genes), or a gene x cell type weight matrix
(marker scores = weighted average expression).}
}
\value{
A cell type x cell matrix containing marker scores.
}
\description{
Compute marker scores (average marker expression) in a target dataset
}
