% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{summarize_auroc}
\alias{summarize_auroc}
\title{Compute separability (AUROC) of cell types based on marker sets.}
\usage{
summarize_auroc(scores, true_labels)
}
\arguments{
\item{scores}{A marker score matrix as returned by score_cells or compute_marker_enrichment.}

\item{true_labels}{Ground truth labels for the target cells.}
}
\value{
A marker set x true label AUROC matrix. For a given combination, the
marker scores are used as a predictor and the true labels define a binary
classification problem (do cells belong to the given cell type?).
}
\description{
Compute separability (AUROC) of cell types based on marker sets.
}
